backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
bcrypt==4.2.0
Pillow==10.1.0
python-dotenv==1.0.0
marshmallow==3.21.0
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1